Move ResponseDataProvider to page_input_providers.py and create a new way to register providers
