+++ scrapy_poet/providers.py
@@ -7,10 +7,10 @@ external sources, for example, Splash or Auto Extract API.
 """
 import typing
 
-from scrapy.http import Response
-from web_poet.page_inputs import ResponseData
-
-from scrapy_poet.page_input_providers import PageObjectInputProvider
+from scrapy_poet.page_input_providers import (
+    PageObjectInputProvider,
+    ResponseDataProvider,
+)
 
 # FIXME: refactor _providers / provides / register,  make a nicer API
 
@@ -18,19 +18,26 @@ providers = {}
 
 
 def register(provider_class: typing.Type[PageObjectInputProvider],
-             provided_class: typing.Type):
+             provided_class: typing.Optional[typing.Type] = None):
     """This method registers a Page Object Input Provider in the providers
     registry. It could be replaced by the use of the ``provides`` decorator.
 
-    Example:
+    If ``provided_class`` is not specified, we get it from ``provider_class``.
 
+    Examples:
+
+        register(ResponseDataProvider)
         register(ResponseDataProvider, ResponseData)
     """
+    if provided_class is None:
+        provided_class = provider_class.provided_class
+
     providers[provided_class] = provider_class
 
 
+# TODO: does this decorator still make sense?
 def provides(provided_class: typing.Type):
-    """This decorator should be used with classes that inherits from
+    """This decorator could be used with classes that inherits from
     ``PageObjectInputProvider`` in order to automatically register them as
     providers.
 
@@ -43,19 +50,4 @@ def provides(provided_class: typing.Type):
     return decorator
 
 
-@provides(ResponseData)
-class ResponseDataProvider(PageObjectInputProvider):
-    """This class provides ``web_poet.page_inputs.ResponseData`` instances."""
-
-    def __init__(self, response: Response):
-        """This class receives a Scrapy ``Response`` as a dependency."""
-        self.response = response
-
-    def __call__(self):
-        """This method builds a ``ResponseData`` instance using a Scrapy
-        ``Response``.
-        """
-        return ResponseData(
-            url=self.response.url,
-            html=self.response.text
-        )
+register(ResponseDataProvider)
